fixed the pack method from DomainAwareDataset
corrected the "bug" making the pack method return floats in the domain_sample arrays instead of integers

+++ skada/datasets/_base.py
@@ -194,7 +194,7 @@ class DomainAwareDataset:
             # xxx(okachaiev): this is horribly inefficient, re-write when API is fixed
             Xs.append(X)
             ys.append(y)
-            sample_domains.append(np.ones_like(y)*domain_id)
+            sample_domains.append(np.ones_like(y).astype(int)*domain_id)
             domain_labels[domain_name] = domain_id
         # xxx(okachaiev): code duplication, re-write when API is fixed
         dtype = None
@@ -224,7 +224,7 @@ class DomainAwareDataset:
             # xxx(okachaiev): this is horribly inefficient, rewrite when API is fixed
             Xs.append(X)
             ys.append(y)
-            sample_domains.append(-1 * domain_id * np.ones_like(y))
+            sample_domains.append(-1 * domain_id * np.ones_like(y).astype(int))
             domain_labels[domain_name] = -1 * domain_id
 
         # xxx(okachaiev): so far this only works if source and target has the same size
